+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 
+## Unreleased
+
+### Added
+- Version (`-version`) CLI flag
+
+
 ## v0.3.0
 
 ### Added